Fix build issue involving missing import
