+++ index.html
@@ -16,8 +16,8 @@
   loadMarkdown("/feed/README.md", "readmeDiv", "_parent");
   </script>
 
-  <script type="module" crossorigin src="./assets/index-7cd7c963.js"></script>
-  <link rel="stylesheet" href="./assets/index-d9535706.css">
+  <script type="module" crossorigin src="./src/feedplayer.js"></script>
+  <link rel="stylesheet" href="./assets/index-388562d5.css">
 </head>
 
 <body>
