+++ widget/assets/index.js
@@ -1,0 +1,1 @@
+(function(n){typeof define=="function"&&define.amd?define(n):n()})(function(){"use strict";(function(){function n(t){var e=document.createElement("script");e.src=t,e.type="module",e.crossOrigin="anonymous",document.head.appendChild(e)}function o(t){var e=document.createElement("link");e.href=t,e.rel="stylesheet",document.head.appendChild(e)}function s(t,e){var i=window.location.pathname;console.log("The current path is "+i),i.includes("/dist/")?console.log("Current path contains /dist/, files will not be loaded."):(n(t),o(e))}s("/feed/dist/assets/index-61ddd40b.js","/feed/dist/assets/index-445cb8a7.css")})()});
